requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
diskcache==5.6.3
grpcio==1.65.4
grpcio-tools==1.65.4
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
Jinja2==3.1.4
llama_cpp_python==0.2.85
macholib==1.16.3
MarkupSafe==2.1.5
numpy==2.0.1
packaging==24.1
pipdeptree==2.23.1
portalocker==2.10.1
protobuf==5.27.3
pydantic==2.8.2
pydantic_core==2.20.1
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
PyYAML==6.0.1
qdrant-client==1.10.1
setuptools==72.1.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.2
